requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
attrs==25.3.0
audioop-lts==0.2.1
beautifulsoup4==4.14.2
bs4==0.0.2
certifi==2025.6.15
charset-normalizer==3.4.2
discord==2.3.2
discord.py==2.5.2
frozenlist==1.7.0
google_search_results==2.4.2
idna==3.10
instaloader==4.14.1
multidict==6.6.0
pillow==11.2.1
propcache==0.3.2
python-dotenv==1.1.1
requests==2.32.4
soupsieve==2.8
typing_extensions==4.15.0
urllib3==2.5.0
yarl==1.20.1
